+++ avk/algorithms/thatsoven_stable_sort.cpp
@@ -5,5 +5,5 @@
 
 void thatsoven_stable_sort(main_array array)
 {
-	thsSort<item>().stableSort(array.begin(), 0, array.size());
+	thsSort<item>().stableSort(array.begin(), 0, (int)array.size());
 }
